Added math.

+++ pong_core/src/constants.rs
@@ -1,6 +1,6 @@
-pub const LCD_HEIGHT: i32 = 360;
-pub const LCD_WIDTH: i32 = 640;
-pub const PADDLE_OFFSET: i32 = 50;
-pub const PADDLE_SPEED: i32 = 1;
-pub const PADDLE_HEIGHT: i32 = 90;
-pub const PADDLE_WIDTH: i32 = 10;
+pub const LCD_HEIGHT: f32 = 360.0;
+pub const LCD_WIDTH: f32 = 640.0;
+pub const PADDLE_OFFSET: f32 = 50.0;
+pub const PADDLE_SPEED: f32 = 1.0;
+pub const PADDLE_HEIGHT: f32 = 90.0;
+pub const PADDLE_WIDTH: f32 = 10.0;
